Remote bitrate estimator (#80)
Habemus REMB

* rbe: initial commit

* rbe: new 'LatesTimestamp' time utility

* rbe: new log tag 'rbe' (Remote Bitrate Estimator)

* rbe: logging and assertion normalization

* rbe: use Mediasoup RtpPacket and ExtensionHeader implementations

* rbe: don't consider RTCP Feedback packets for now. TODO

* rbe: use internal RtpDataCounters

* rbe: set a fixed minimum bit rate

* rbe: remove adaptive threshold experiment

...until we understand the adaptive threshold parameters

http://c3lab.poliba.it/images/6/65/Gcc-analysis.pdf (section 5)

* rbe: add files into .gyp file

* add 'goog-remb' to the supported RTP capabilties

* rbe: rename 'RemoteBitrateObserver' to 'RemoteBitrateEstimator::Listener'

* rbe: make use of the remote bitrate estimator

* rtcp: REMB packet

* remb: send RTCP REMB messages when remote bitrate calculator triggers a change

* remb: cosmetic

* remb: update doc

* rbe: cosmetic

* rtcp: make REMB packet a specialization of AFB

* rbe: remove unused RemoteBitrateEstimator::IncomingPacketFeedbackVector()

Note: also removed in original implementation

* RateCalculator: new Reset() interface

* rbe: Mediasoup-ization first commit

get rid of BweDefines.hpp

* prepend 'this->' to instance members'

* rbe: prepend data members with 'this->' and camelCase them

* rbe: camelCase local variables and method arguments

* rbe: use tabs and inline small and simple member methods

* rbe: remote ctags file

* rbe: mayor style changes

* rbe: add MS_TRACE macro

* rbe: use 'default' keyword for empty desructor on base class

* rbe: more cleanup

* rbe: declare static data before any other

* rbe: start comments with lowercase

* rbe: always use adaptive thresholds

* include missing header file

* remb: make getters const

* Update REMB stuff to new RTP extension API

* remb: use specific API to get abs exten.

* Clean up how REMB, NACK and RTP header extensions are checked.

+++ worker/include/RTC/RtpDictionaries.hpp
@@ -194,7 +194,7 @@ namespace RTC
 		Media::Kind                 kind = Media::Kind::ALL;
 		std::string                 uri;
 		RtpHeaderExtensionUri::Type type;
-		uint16_t                    preferredId = 0;
+		uint8_t                     preferredId = 0;
 		bool                        preferredEncrypt = false;
 	};
 
@@ -276,7 +276,7 @@ namespace RTC
 	public:
 		std::string                 uri;
 		RtpHeaderExtensionUri::Type type;
-		uint16_t                    id = 0;
+		uint8_t                     id = 0;
 		bool                        encrypt = false;
 		RTC::Parameters             parameters;
 	};
